fix(force-run-modal): move to services domain
