+++ libs/domains/services/feature/src/index.ts
@@ -61,6 +61,7 @@ export * from './lib/service-overview/service-overview'
 export * from './lib/service-overview/observability-callout'
 export * from './lib/service-state-chip/service-state-chip'
 export * from './lib/service-clone-modal/service-clone-modal'
+export * from './lib/force-run-modal/force-run-modal'
 export * from './lib/service-terminal/service-terminal'
 export * from './lib/need-redeploy-flag/need-redeploy-flag'
 export * from './lib/hooks/use-clean-failed-jobs/use-clean-failed-jobs'
