+++ libs/shared/console-shared/src/index.ts
@@ -1,7 +1,5 @@
 // This library is deprecated and will be removed in the future
 // Please use util or domain libraries instead
-export * from './lib/force-run-modal/feature/force-run-modal-feature'
-export * from './lib/force-run-modal/ui/force-run-modal'
 export * from './lib/entrypoint-cmd-inputs/ui/entrypoint-cmd-inputs'
 export * from './lib/flow-create-port/ui/flow-create-port/flow-create-port'
 export * from './lib/flow-create-port/ui/crud-modal/crud-modal'
